Add caching to tex plugin

+++ _plugins/tex.rb
@@ -12,25 +12,41 @@ module Jekyll
     end
 
     def convert(content)
-      # Temporarily write the content to a temporary file
-      temp_input_path = File.join(Dir.tmpdir, "#{SecureRandom.uuid}.tex")
-      File.write(temp_input_path, content)
+      # Calculate a unique hash for the content of the .tex file
+      content_hash = Digest::MD5.hexdigest(content)
 
-      # Define the Docker command to run engrafo and cat the HTML output
-      docker_command = "sudo docker run -v #{Shellwords.escape(temp_input_path)}:/workdir/input.tex -w /workdir arxivvanity/engrafo sh -c 'engrafo input.tex /tmp --javascript /assets/engrafo.js --css \"data:text/css;base64,\" >&2 && cat /tmp/index.html'"
+      # Check if there is a cached HTML version for this content
+      cached_html_path = File.join(Dir.tmpdir, "tex_converter_cache_#{content_hash}.html")
 
-      # Execute the Docker command and capture the HTML output
-      html_output = `#{docker_command}`
+      if File.exist?(cached_html_path)
+        # If a cached version exists, return it
+        return File.read(cached_html_path)
+      else
+        # If no cached version exists, proceed with the conversion
 
-      # Check for errors in the Docker command execution
-      unless $?.success?
-        raise "Error running engrafo:\n#{html_output}"
+        # Temporarily write the content to a temporary file
+        temp_input_path = File.join(Dir.tmpdir, "#{SecureRandom.uuid}.tex")
+        File.write(temp_input_path, content)
+
+        # Define the Docker command to run engrafo and cat the HTML output
+        docker_command = "sudo docker run -v #{Shellwords.escape(temp_input_path)}:/workdir/input.tex -w /workdir arxivvanity/engrafo sh -c 'engrafo input.tex /tmp --javascript /assets/engrafo.js --css \"data:text/css;base64,\" >&2 && cat /tmp/index.html'"
+
+        # Execute the Docker command and capture the HTML output
+        html_output = `#{docker_command}`
+
+        # Check for errors in the Docker command execution
+        unless $?.success?
+          raise "Error running engrafo:\n#{html_output}"
+        end
+
+        # Clean up the temporary input file
+        File.delete(temp_input_path)
+
+        # Save the generated HTML to the cache for future use
+        File.write(cached_html_path, html_output)
+
+        html_output
       end
-
-      # Clean up the temporary input file
-      File.delete(temp_input_path)
-
-      html_output
     end
   end
 end
